requirements.txt
Babel==0.9.6
Beaker==1.7.0
FormEncode==1.2.6
Genshi==0.6
Jinja2==2.6
Mako==0.9.0
MarkupSafe==0.18
Pairtree==0.7.1-T
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
Pygments==1.6
Pylons==0.9.7
Routes==1.13
SQLAlchemy==0.9.6
Tempita==0.5.2
WebError==0.10.3
WebHelpers==1.3
WebOb==1.0.8
WebTest==1.4.3
argparse==1.2.1
ckanext-google-tag-manager==0.0.1
decorator==3.4.0
fanstatic==0.12
nose==1.3.0
ofs==0.4.1
passlib==1.6.2
pbr==0.8.2
psycopg2==2.4.5
python-dateutil==1.5
pyutilib.component.core==4.5.3
repoze.lru==0.6
repoze.who==2.0
repoze.who-friendlyform==1.0.8
requests==2.3.0
simplejson==3.3.1
six==1.7.3
solrpy==0.9.5
sqlalchemy-migrate==0.9.1
sqlparse==0.1.11
unicodecsv==0.9.4
vdm==0.13
wsgiref==0.1.2
zope.interface==4.1.1
